requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
facenet-pytorch==2.6.0
filelock==3.15.1
fsspec==2024.6.0
idna==3.7
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
opencv-contrib-python==4.10.0.82
opencv-python==4.10.0.82
pillow==10.2.0
requests==2.32.3
sympy==1.12.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.1
